+++ Libs/LibClassicDurations/CHANGELOG.md
@@ -1,6 +1,9 @@
 # LibClassicDurations
 
-## [1.19](https://github.com/rgd87/LibClassicDurations/tree/1.19) (2019-09-13)
-[Full Changelog](https://github.com/rgd87/LibClassicDurations/compare/1.18...1.19)
+## [1.44](https://github.com/rgd87/LibClassicDurations/tree/1.44) (2019-12-26)
+[Full Changelog](https://github.com/rgd87/LibClassicDurations/compare/1.41...1.44)
 
-- Remember Rupture and Kidney Shot calulcated duration  
+- Don't keep buff cache status table between library instances  
+- Enemy buffs cache only valid for 40s  
+- SnD plug  
+- Returning buff UnitAura values up to timeMod